added `nu-utils` crate, fixed issue where externals turn off vt processing
