+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
 	"crates/nu_plugin_gstat",
 	"crates/nu_plugin_example",
 	"crates/nu_plugin_query",
+	"crates/nu-utils",
 ]
 
 [dependencies]
@@ -50,6 +51,7 @@ nu-protocol = { path = "./crates/nu-protocol", version = "0.59.1"  }
 nu-system = { path = "./crates/nu-system", version = "0.59.1" }
 nu-table = { path = "./crates/nu-table", version = "0.59.1"  }
 nu-term-grid = { path = "./crates/nu-term-grid", version = "0.59.1"  }
+nu-utils = { path = "./crates/nu-utils", version = "0.59.1" }
 pretty_env_logger = "0.4.0"
 rayon = "1.5.1"
 reedline = { git = "https://github.com/nushell/reedline", branch = "main" }
@@ -90,7 +92,7 @@ opt-level = "s" # Optimize for size
 strip = "debuginfo"
 lto = "thin"
 
-# build with `cargo build --profile profiling` 
+# build with `cargo build --profile profiling`
 # to analyze performance with tooling like linux perf
 [profile.profiling]
 inherits = "release"
