+++ crates/nu-cli/Cargo.toml
@@ -8,7 +8,7 @@ nu-engine = { path = "../nu-engine", version = "0.59.1"  }
 nu-path = { path = "../nu-path", version = "0.59.1"  }
 nu-parser = { path = "../nu-parser", version = "0.59.1"  }
 nu-protocol = { path = "../nu-protocol", version = "0.59.1"  }
-# nu-ansi-term = { path = "../nu-ansi-term", version = "0.59.1"  }
+nu-utils = { path = "../nu-utils", version = "0.59.1"  }
 nu-ansi-term = "0.42.0"
 
 nu-color-config = { path = "../nu-color-config" }
