+++ crates/nu-command/Cargo.toml
@@ -23,6 +23,7 @@ nu-system = { path = "../nu-system", version = "0.59.1"  }
 nu-table = { path = "../nu-table", version = "0.59.1"  }
 nu-term-grid = { path = "../nu-term-grid", version = "0.59.1"  }
 nu-test-support = { path = "../nu-test-support", version = "0.59.1"  }
+nu-utils = { path = "../nu-utils", version = "0.59.1" }
 
 # Potential dependencies for extras
 base64 = "0.13.0"
