+++ crates/nu-utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+authors = ["The Nushell Project Developers"]
+description = "Nushell utility functions"
+edition = "2021"
+license = "MIT"
+name = "nu-utils"
+version = "0.59.1"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[[bin]]
+name = "utils"
+path = "src/main.rs"
+
+[dependencies]
+
+[target.'cfg(windows)'.dependencies]
+crossterm_winapi = "0.9.0"
