+++ crates/nu-utils/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod utils;
+
+pub use utils::enable_vt_processing;
